Restructure project as proper Python package
- Move all Python files into stufflog package directory
- Create __init__.py with package metadata
- Update all imports to use absolute paths
- Fix setup.py to use proper package structure
