+++ stufflog/__init__.py
@@ -1,0 +1,7 @@
+"""
+stufflog - A general journalling CLI application that uses YAML files.
+"""
+from .stufflog import main
+
+__version__ = "0.1.0"
+
