+++ stufflog/exceptions.py
